==> data_processing/01d-cleanOrgWeather_Radiation.R <==
# This code organizes data from the quantum sensor logger.
# It deals with broken text files generated by the attempt of
# automating extraction of data from the logger.

# Set configurations ------------------------------------------------------
rm(list = ls())
options(stringsAsFactors = FALSE)
Sys.setenv(LANG = "En")

# Libraries ---------------------------------------------------------------


# Functions ---------------------------------------------------------------

source("./data_processing/01d-cleanOrgWeather_Radiation_functions.R")

orgSensorData <- function(x){
  
  x_mod <- x %>%
    mutate(measurement= if_else(measurement < 0, 0, measurement)) %>%
    formatData() %>%
    cleanDup() %>%
    fillNight()
  
  return(x_mod)
  
}


# Load files --------------------------------------------------------------

rad_files <- c(list.files("./data/weather/monitoring/raw/li1400/", full.names = T, 
                        pattern = "*.txt"),
               list.files("./data/weather/monitoring/raw/li1400/01-manually_cleaned/", 
                          full.names = T, 
                          pattern = "*.txt"))

if (file.exists("./data/weather/monitoring/raw/rad_raw.csv")){
  
  unlink("./data/weather/monitoring/raw/rad_raw.csv")
  
}

for (i in rad_files){
  
  temp <- readLines(i,warn = FALSE) #Read whole file
  temp <- data.frame(all = temp[(grep("^1\t", x = temp))]) #Filter measurement lines
  temp_df <- separate(temp, 
                      col = "all",
                      into = c("code", "dateFull", "rad1", "rad2"),
                      sep = "\t") %>%
    select(-code) %>%
    gather('rad1', 'rad2', key = 'node', value = 'measurement') %>%
    mutate(measurement = str_replace_all(measurement, "\\^M", "")) %>%
    drop_na() %>%
    mutate(node = substring(node, 4, 4),
           measurement = as.numeric(measurement)) %>%
    filter(!is.na(measurement)) 
  
  #print(summary(temp_df))
  write.table(temp_df, file = "./data/weather/monitoring/raw/rad_raw.csv",
              sep = ",",
              append = TRUE,
              row.names = FALSE,
              col.names = FALSE,
              fileEncoding = "UTF-8")
  
}

head_csv <- c("dateFull", "node", "measurement")

rad <- read.csv("./data/weather/monitoring/raw/rad_raw.csv", dec = ".",
                header = FALSE)
colnames(rad) <- head_csv

# Process data ------------------------------------------------------------

rad_full <- orgSensorData(rad)
write.csv(rad_full, 
          file = "./data/weather/monitoring/weather_radiation_org.csv", 
          row.names=FALSE)
